refactor to one Util class
